google-adk-agents/weather-bot/requirements.txt
google-adk
litellm
# For CLI and agent tools
# No external weather API required (mocked)
